Detect environment font when rendering canvas
We discovered that the fonts used for the perf panel canvas are hard
coded, but this causes issues because for screenshot tests we need to
enforce a specific font. This CL creates a helper and updates the canvas
code to use it. It also removes some font properties on Group styles in
the Perf Panel, because we never used them to change or override the
font.

Bug: none
Change-Id: I461ebb8ab8ba445f2620f9fa79f18442151b5994