Remove uiSourceCodes from NetworkPersistenceManager on project removal
The `NetworkPersistenceManager` contains
`networkUISourceCodeForEncodedPath`, a map from paths to
`UISourceCodes`. Scripts are not properly cleared from this map upon
page navigation and accumulate in the map, whereas other types of
resources are removed upon navigation.
This CL ensures that when a project is removed, its `UISourceCodes`
are removed from `networkUISourceCodeForEncodedPath`.

Fixed: 1422576
Change-Id: I1df732b74ddc3feacc6e54cbf3a0c40b3291858b