Export merged results of unit and interaction test coverage as text file
For convenience, since the bots will still use the JSON file.

Also, an option to define the folders on which coverage will be
calculated is added to the interaction tests. The flag works
identically to the COVERAGE_FOLDERS of unit tests, i.e. a pattern is
provided to define the included folders. For example: the flag can be
set like COVERAGE_FOLDERS=
front_end/{models/trace,models/timeline_model,panels/timeline}

This two changes should help to calculate the total OPP coverage in
go/opp-test-coverage

Bug: 1424729
Change-Id: Ie982f62dfac0014c05075e53173a8e7710d3b9e0