Ensures screen reader announces the correct state of checkboxes inside of categorizedBreakpointSideBarPane.ts
There was some latency between when the screen reader picks up the aria-checked property and when the updated aria-checked properties are set.

This was because when a breakpoint is selected, it gains focus, so the screen reader picks up the incorrect aria-checked property just before the click handler sets the aria-checked attribute.

This patch fixes this by setting focus on the checkbox itself when the listItemElement is selected. WIth this, the issue around the element label being announced twice is also resolved.

To ensure keyboard navigation still works after the checkbox is focused, a key event listener is attached to the checkbox to refocus the listItemElement in order to allow the treeElement key event listener to run normally without halting abruptly due to the isEditing check.


Bug: 1414952
Change-Id: I8d33c1f4a04cf0ac1ef5d6e8da4605e68890f35c