Accessible names for color buttons in CSS Overview
This change ensures screen readers have descriptive text for buttons
appearing under the Colors section of a completed CSS Overview.

Color buttons have a sibling div element with the hex color text.
The div also had a title attribute. The div is not the element
that gets focus though, so this title attribute was not being used for
screen readers. The fix is to move the title attribute to the button
so that the title is announced when the user focuses the button.

Contrast issue buttons have a descriptive title attribute, but in
Narrator on Windows, this attribute was superceded by the button's
child text, which is simply "Text". The fix is to add an aria-label
alongside title which takes precedence over the child text ensuring
the intended description will be read by Narrator.

Bug: 1425590
Change-Id: I842653e589d64f005caed62ac64090b625e60035