[console] Add setting to accept autocomplete suggestion on Enter.
This behaves similar to the "Accept Suggestion On Enter" setting in VS
Code, and is also disabled by default just like in VS Code. When
disabled the default conservative autocompletion will be used in the
Console, which only accepts the suggestion on Enter after explicitly
interacting with the completion popup, and otherwise only allows to
accept the suggestion via Tab or ArrowRight.

When enabled, the autocompletion behaves just like the regular one in
the Sources panel, such that Enter will always accept the suggestion.

Bug: chromium:1276960
Change-Id: Ia9f295321fd6b103cd0f2bafad3e962a3157118a
Fixed: chromium:1425722
Screenshot: https://imgur.com/XxotS9Z.png