Support conflicting sources in source maps
This change fixes the case where there are multiple source maps
that contain different versions of the same file. It fixes faulty
logic in SourceFrame.ts to ensure we always display the code when
a new UISourceCode is set. It also fixes CompilerScriptMapping to
create a project per source map URL so that we can handle conflicting
source maps even when they are in the same worker/frame.

Fixed: 1111297
Change-Id: Ib19b7b9ad6095020b3c6186d7ac91a36888d4218