[Header Overrides] Save changes on 'focus out' in sources panel editor
Until now, when editing header overrides in the sources panel, the
user had to save their changes manually via a keyboard shortcut. In
the network panel, changes to header overrides are saved as soon as
the edited element loses focus.
This CL creates consistent behaviour by also saving changes in the
sources panel as soon as a field loses focus.

Fixed: 1416334
Change-Id: If3c6d714bff7b4264527d625d503bfdf1f6b6543