[RPP] Use a static script url for JS call frames
Before this CL the script url was a live location tied to the script's
id. This means that the url of a call frame will be the url of the
script with id equal to the callframe's script id. This is problematic
when the inspected target is not the same as the one a trace was taken
from or when a trace is loaded from a file: If a script id is found
that matches the script id of a call frame in the trace, that script's
url will be assigned to the call frame, even if the script is not the
one the call frame's scriptid points to.

To fix this, the displayed url is taken from the static
Linkifier.linkifyURL function, which uses the script url in the
call frame and the url of the inspected target (to determine if the
origin of the script is shown), and is also independent from the script
id. With this the url should always point to the same file.

Fixed: 1425523
Change-Id: Ib8642ec4f89fa14161053bfb132fcb181eea01a8