[RPP] Add support for console timings in the UserTimingsHandler
These are the events currently shown by the OPP in the Console track,
with the addition of events triggered by the console.timeStamp API
which, at the moment, are not shown.

Because console timing events are built with trace event begin and end
tuples, just like performance.measures, the parsing algo and some types
related to performance.measures are extended and reused. Dedicated types
for console timings and for performance measures are introduced as well.

Bug: 1425422
Change-Id: If7f20b5a9622890b3d84b46a62da89233f7df664