[settings] Add separate setting version number for each storage  bucket
This CL adds 2 new settings that represent the settings version for
the synced and local storage bucket respectively.

We take the minimum version of the 3 and run all the migrations from
that minimum up to the current version.

This CL also adds a couple of unit test that checks that the
correct update methods are called in various scenarios.

R=jarin@chromium.org

Bug: 1426748
Change-Id: Idc705bad12bdf5a631f07515f901c7e82a061195