Check if manually attached sourcemaps appear to match the file
When manually attaching sourcemaps to a source file, it's hard to
determine when sourcemaps are incompatible. Since all mappings are bogus
the most likely symptom is entirely erratic behavior when pausing,
stepping, or setting breakpoints.

This CL tries to detect such cases on a best-effort basis by comparing
the sourcemap's `file` property (if present) and the last mapping with
the script it's attached to.


Bug: none-for-now
Change-Id: Ie845d360abb369af41e165e64ab1759f48db3cba