[RPP] Improve Track.syncEvents implementation
The work on crbug.com/1416533 has revealed the relationship between
flame chart tracks and the details pane of the panel as layout tests
complain as tests land. I've noticed that this particular method is a
common dependency among the tabs in the details pane. In fact this is
the data source for the pane and its implementation is a bit confusing.

For example, despite being called "syncEvents" the method can return
data from *async* events. For this reason this CL adds some comments,
tests and changes the name for clarity.

This CL also also makes it possible to return async and sync events at
the same time, when it makes sense. Before, only sync events were
returned if available and this caused information loss when a track had
both async and sync events (e.g. the timings track).

Bug: 1386091


[RPP] Add support for console timings in the UserTimingsHandler

These are the events currently shown by the OPP in the Console track,
with the addition of events triggered by the console.timeStamp API
which, at the moment, are not shown.

Because console timing events are built with trace event begin and end
tuples, just like performance.measures, the parsing algo and some types
related to performance.measures are extended and reused. Dedicated types
for console timings and for performance measures are introduced as well.

Bug: 1425422


[RPP] Add support for console timings in the UserTimingsHandler

These are the events currently shown by the OPP in the Console track,
with the addition of events triggered by the console.timeStamp API
which, at the moment, are not shown.

Because console timing events are built with trace event begin and end
tuples, just like performance.measures, the parsing algo and some types
related to performance.measures are extended and reused. Dedicated types
for console timings and for performance measures are introduced as well.

Bug: 1425422
Change-Id: Iebdeec134c8389b877f591318cf40b760c7ab167