[RPP] Improve Track.syncEvents implementation
The work on crbug.com/1416533 has revealed the relationship between
flame chart tracks and the details pane of the panel as layout tests
complain as CLs land. I've noticed that this particular method is a
common dependency among the tabs in the details pane. In fact this is
the data source for the pane and its implementation is a bit confusing.

For example, despite being called "syncEvents" the method can return
data from *async* events. For this reason this CL adds some comments,
tests and changes the name for clarity.

This CL also also makes it possible to return async and sync events at
the same time, when it makes sense. Before, only sync events were
returned if available and this caused information loss when a track had
both async and sync events (e.g. the timings track).

Bug: 1386091
Change-Id: Iebdeec134c8389b877f591318cf40b760c7ab167