Add Storage Buckets to Application panel
Adds the Storage Buckets menu to the sidebar of the Application panel.
The Storage Buckets menu is a tree which contains a list of all the
storage buckets for the page. Under each storage bucket will be  a list
of supported storage APIs. To use, this must be enabled in the
Devtools' experiments settings menu.

Bug: 1406017
Change-Id: I66840c0f12624486c61e15775ebbfb13d8f38c7d