Update goldens even if threshold is met when force update flag is set
Before this CL setting the flag meant that screenshots would be updated
only if the diff that was obtained in a test was above the defined
threshold. So that screenshots are more likely to be kept updated this
CL changes this behaviour so that if the flag is set, any image diff >0
would trigger an update.

Bug: none


Rebaseline mac goldens:

They are failing in CI:
https://ci.chromium.org/ui/p/devtools-frontend/builders/ci/Stand-alone%20Mac/9878/overview

Bug: none
Change-Id: Ic7af68589d09dd6082aeb5237b3dbb9c4fd138f6