Fix problem where guideline fail to show up
a liner gradient has no natural size and its concrete size will
match the size of the element it applies to. If the color part of
the gradient is too small in relation to the container,
then the gradient applied to a long wide container may not be visible.
https://codepen.io/koheiueno/pen/bGxZEEd

How to test:
  1. Open https://google.com
  2. Open Chrome DevTools
  3. Open the Elements panel
  4. Expand the body element
  5. See if guidelines are visible

Before:  https://imgur.com/a/I4Tkyb4
After:  https://imgur.com/a/sYWmRtz
Bug: 1424968
Change-Id: I108fd3ca69cb99d347aeada04812f61a0ab73356