[sources] Update code folding icons to GM3.
This imports the Material 3 versions of triangle-down.svg and
triangle-right.svg, and updates the size of the folding markers
(unusual icon size case, matching the 14x14 dimensions that the
folding markers in the Elements panel are supposed to have). Also
adds a new color token for the fold markers.

Bug: chromium:1427397
Screenshot: https://imgur.com/KedX0Ay.png
Change-Id: I4cdb8bc62698fba9dc8e5d30228622a1bae00c2a