[sources] Switch debugger actions to GM3 icons.
This updates the debugger actions to use the GM3 icons (also switching
the AnimationTimeline as drive-by since that's using the pause and
resume icons as well). It does not switch the disable breakpoints icon
yet, since that is also changing how it's displayed generally.

Screenshot: https://imgur.com/eTuiUn7.png
Bug: chromium:1427397
Change-Id: I454b9a37e5c0d2893171eaee4b3d4eb22d46dbdb