Autostep over unmapped wasm locations
Upon step over starting from a mapped wasm location, devtools will automatically continue stepping if the step causes us to pause on an unmapped wasm location. Autostepping also ends when it lands outside of the function we started in, i.e., after we returned from the function. The behavior is controlled by a new setting which is on by default.

Fixed: 1418938
Change-Id: I45010bc6536b01c2d00dba40348c585c08672a4e