[debugger] Fix execution line leak with same URL.
Previously we'd key the execution line highlighting on the URL and
highlight the execution in all editors whose UISourceCode has the
same URL as the current execution line. This made sense when we'd
deduplicate only on URL and only ever showed a canonical source (in
the hope that this canonical source is representative of all the
sources with the same URL).

However with the recent changes to always reveal the actual source
in the deduplicated tab (specifically https://crrev.com/c/4231795),
we can just limit the execution line highlighting the actual source
since we no longer need to deal with the case that we're showing a
canonical source instead.

Bug: chromium:1402190
Change-Id: I887051da3a8e739e2fa9b340c84a75127a6cf023
Fixed: chromium:1427436