Enable translation of ViewLocationCategoryValues into other languages
Strings in ViewLocationCategoryValues needed to be used as identifiers
to do comparisons but also as text to be displayed to the user in other
languages. This creates a problem where the calling code needs to be
aware of the localized value.

This CL fixes the problem by using an enum and a method that provides
a translated UI string given a ViewLocationCategory.

R=szuend@chromium.org

Bug: chromium:1181019
Change-Id: Ibb22ab7fa1cf540da07734c9ca581a82a0c8be80