[console] Fix copying text from structured console messages
This CL fixes a regression introduced by https://crrev.com/c/4241807.
When the user copies text from the console view, DevTools carefully
massages the selection by traversing the DOM and ignoring certain
nodes.

When we introduced devtools-icon (a custom element), we needed to
change the traversion code as otherwise we'd include our own
HTML code comments from the Lit templates. The above CL was not
complete and we missed one variance of this logic in the
ConsoleViewPort.

R=pfaffe@chromium.org

Fixed: 1428966
Change-Id: I2d6fe4c86cf7927c62dfed0df7549f89d33bd7bc