Enable scoring panels and drawers in quick menu for other languages
The action 'category' can be a translated string, but currently we
explicitly check if the category starts with either the 'Panel' or
'Drawer', and give those commands a higher score so they appear higher
in the list. This only works for en-US DevTools.

This CL introduce an option isPanelOrDrawer in the CreateCommandOptions
interface and the Command class along with a const enum PanelOrDrawer
so that we can check the category using this option and calculate
the scoring as usual.

R=szuend@chromium.org

Bug: chromium:1429478
Change-Id: I221bb9d915988bd5100034a63235d7c835e2f114