Animate tree-item expansion
The animation of the triangle icon when clicking on an issue was
broken. This CL fixes this and instead of only animating tree item
triangles in the issues panel, it makes the CSS animation apply to
all legacy trees.

Video: https://imgur.com/a/3dBTne0

Mask sizes and mask positions are not needed anymore because the new
icon files contain only a single icon each.

Drive-by: fix comment line whitespace in issuesTree.css

Bug: 1427397
Change-Id: I6fa09f5a31a12d2776a8ae999f04a820e891f6cb