[icon] Update the sub-menu icon for SoftContextMenu to GM3
Note: To take the screenshots I locally changed the context menu in
the 'Elements' panel to 'soft' instead of using the native one.
By default most DevTools menus are native (unless in hosted mode), so
it's hard to find a use-site for the icon.

Screenshots: https://imgur.com/a/9CTsRZ9

R=kprokopenko@chromium.org

Bug: 1427397
Change-Id: I168186b1cac88efde7bf32b0fb93af098c3e49d1