Unify creating animation timing function models and their rendering
The main idea is, we want to be able to render `cubic bezier` visualization for the `cubic-bezier()`functions and `linear` visualization for `linear()` functions.

For this, I have:
* Created `AnimationTimingModel.ts` to be used as the common interface
to interact with the underlying model regardless of it being a UI.Geometry.CubicBezier or CSSLinearEasingModel.

The only things we need there in the outside is to "be able to convert
text to an `AnimationTimingModel` and "be able to get CSS text from the same model".

In the places of `UI.Geometry.CubicBezier.parse` I have used `AnimationTimingModel.parse` which will also include parsing `CSSLinearEasingModel` as the next step.

* Created `AnimationTimingUI` to be used as the common "component" for rendering the visualization for the models and communicating with the BezierEditor. I didn't want to change how BezierUI and BezierEditor is currently interacts with each other: BezierEditor provides a "container" div and BezierUI renders itself to this div. Though, the instance for the `BezierUI` must be kept around so that redraws should not append more divs to the container.

Because of this, I have added a logic in `AnimationTimingUI` that creates one "BezierUI" whenever it is needed and keeps it around. It will be similarly with the LinearEasingUI tool.

Bug: 1421241
Change-Id: I7d4f1e05f8ff290b201d9301941c47b342663490