Fix TimelineSelection with new appender events
I naively enabled TimelineSelection classes to take new trace events,
and thought because TS compiled, it would work :D

Turns out that was foolish; we cast the result of selection.object() in
multiple places, so I didn't catch many places where I needed to deal
with an event not being an SDK event.

I will continue to figure out how to deal with that, but in the mean
time this CL reverts the part of the change to ensure we always pass in
actual SDK events, which I have verified fix the problem.

Bug: 1434291
Change-Id: I22d1ff59a09f5fa8d71ba4fe3542fa64ee576d66