[RPP] Support new engine trace event type in some helpers in TimelineUIUtils
I chose these helpers because either the support is easily added or is
needed when trying to support new engine types in the details pane
artifacts.

Note there are new early returns in case of SDK.Event occurrences. This
are included because the feature that is skipped by each of these checks
is not implemented by the new engine. As new features are added, we can
come back to these checks and use the new engine appropriately.

Bug: 1433692
Change-Id: I9290cbd7fa599cbb447b7cd0d7cec05601280708