[icon] Fix SoftContextMenu alignment regression
The CL https://crrev.com/c/4422328 changed the icon implementation
in the SoftContextMenu to <devtools-icon>. This broke alignment when
the menu item text was long enough to cause the menu to grow.

The root cause is that <devtools-icon> when hidden, reports a width
of 0, throwing off the size calculations. This CL fixes this by
setting `min-width` on the hidden checkmark icon.

R=kprokopenko@chromium.org

Bug: 1427397
Change-Id: I785494e775bf5a8b2679ab1092dae682dafb4178
Fixed: 1434260