[web test] Skip <devtools-icon> when serializing DOM
All CSS tests traverse the style panel DOM and serialize it for
test expectations. We need to skip the <devtools-icon> in this logic,
as otherwise we include the formatting whitespace of the Lit templates.

We suspect that the reason why the tests failed flakily and not
consistently is because <devtools-icon> uses the RenderCoordinator,
which renders via `requestAnimationFrame`, so timing is not fixed.

R=jarin@chromium.org

Bug: 1434510,1434632,1434466,1434481,1434731
Change-Id: I39fe1b57fa051dc5828bbc9436b2eb8935b96d9f