Change styles of interaction track for more clarity
As of https://crrev.com/c/4424947 events in interactions tack are not
rendered if they can be nested. This means that's more likely that the
events will only occupy one row in the track. However, the way this
track was programmed before this CL meant that the track was marked as
collapsible even if there is only one row in the track and in this
cases, when the track was expanded, no reaction from the track would be
visible. This is a bit confusing and the user might interpret it as a
bug.

For better user experience, this CL makes sure that when the user
expands the track, the whole track is moved downwards, even if there's
only one row with events, so that the "expanded state" is clear to the
user. This is achieved by teaking the flame chart group to not use
`useFirstLineForOverview`.

Bug: 1432510
Change-Id: Icc1e5b2552fef1b7c76d6af2b8e62fddc85be1a4