Make TimelineSelection type safe
Rewrites TimelineSelection to enforce type-safety, and adds helpers to
assert the instance is a particular type, rather than tracking a
separate enum and objects also.

This means we can avoid `.object()` as Foo, which easily can hide
potential bugs as TS does not check it as you might expect.

Fixed: 1434603
Change-Id: Id2bb790d69f033063f5b4c728ccc16e58d3626cf