Move EventTypeHelpers utils to timeline_model
This way we can also use it in the timeline_model and SDK module without causing circular deps. At first I was hesitant to add it into the legacy folder but I now think it makes sense given that it's added for compatibility between the new engine and the legacy artifacts and once we can remove the legacy artifacts we should also be able to remove the helpers.

Bug: 1433692
Change-Id: I764cb1cce8575076bd9dc5f29cd3f612a308f561