[RPP] Remove `ordinal` field from SDK.TracingModel.Event
This field is synthetically created in the panel for building JS
profiles. Because we don't have such fields in the new engine yet,
references to that field are blocking us for supporting the new engine
types in multiple places. For this reason, this CL removes such field
and adds helpers that support both legacy and new types for keeping
track of the ordinal. This way, references to the ordinal field can be
updated to support new types correspondingly.

Bug: 1433692
Change-Id: I8484726ba43eb1488a3aee5ad7693c903c03069a