Add entryDecorations to FlameChart.ts
This CL adds a definition for event decorations, which can be set in
the FlameChartTimelineData class and will be used as a generic way for
the FlameChart to add decorations to events.

Note that this CL does NOT use this new mechanism, but that will come in
follow-up CLs.

By doing this, we can move the logic for candystriping from the
FlameChart into a data provider, meaning that the FlameChart can remain
generic (right now it looks for main thread events + interactions to add
candy striping too).

I also added a screenshot test that does use the new decoration system
so we have test coverage of that.

Bug: 1434297
Change-Id: I5d9d79ed708f4290988d37b27cca445c6bdf95e3