Use FlameChartDecoration for candy-striping long interactions
This CL updates the candy striping for long interactions to use the new
decoration system.

I will also send another CL that moves the long-task candy striping onto
this system, rather than hardcoding it in the FlameChart.ts component,
but I am purposefully keepign them apart for CL size reasons :)

We already have a screenshot test for candy-striped long interactions,
so we can rely on that test to tell us if this migration has visually
changed anything (it should not have).

Bug: 1434297
Change-Id: I0644e64e9c12c9a7e2a170c0787025f9454ef9e0