[persistence] Don't use CharacterIdMap for FilePathIndex
This CL removes the PathEncoder from FilePathIndex used for detecting
similar files for workspaces. The PathEncoder has a limit to how many
path parts it can hold (65k) and we have encountered crashes in the
wild that bumped into this limit.

Instead of encoding each path part with a codepoint and adding that
codepoint to a Trie, we add each path segment directly to the Trie.

R=bmeurer@chromium.org

Bug: 1436039
Change-Id: I3a7c87cca994f446a1c3981bacc7cfd8efaf1121