[persistence] Remove CharacterIdMap usage from FolderIndex
This CL removes PathEncoder usage from FolderIndex. It has a 65k limit
on how many encoced path segments it can see in total. We use the path
segments themselves directly in the Trie.

Note that FolderIndex balances add/removes call for the same path.
This means of "foo/bar" is added twice, callers need to remove it
twice to purge it fully from the index. This CL keeps this behavior.

R=bmeurer@chromium.org

Bug: 1436039
Change-Id: I6fb3a621c41970bdbfaf173a0abb4685027f5038