[RPP] Use new engine data in the details pane
Instead of resorting to storing the data in legacy tracks. This allows
us to get rid of a bunch of duplicated processing that was there only
because the source of the details pane was the result of the parsing
done in TimelineModel

This is the implementation of the proposal doc to use new engine data in
the details view [1], except we are not forwarding legacy events from
the appender system, we are forwarding raw trace events with the new
engine types because support for these has recently been added throuout
the details view artifacts.

[1] https://docs.google.com/document/d/1yhOwUj0abz6gkc5iDd0YwweO68P0XxX0etgnd0XI54M/edit#heading=h.pcawn2wj7knb

Bug: 1433692
Change-Id: I83659818bd4aaec4e481f250b40c2e1d5b29e41e