Show V8.FinalizeDeserialization time as profiler overhead
The profiler often shows very long blocks of "Compile Script" time, but
much of that time is spent setting up data structures used by the
profiler itself. This setup happens during the event named
V8.FinalizeDeserialization. When no profilers or debuggers are attached,
and no logging is enabled, V8.FinalizeDeserialization does nothing. If
we represent this time as "Profiler Overhead" in the timeline, then it
will be more clear that the "Compile Script" time shown is much larger
than what a normal user would experience.

Bug: chromium:1428605
Change-Id: I6574802568abd4e47dd82eaf6b53cd6a26ef996f