[elements] Hide search on navigation/reload
When redrawing the elements panel, e.g., on reload and navigation, this
CL closes the search toolbar in case it's open instead of only resetting
it. The user can bring it back with single ctrl+f and it will still
contain the previous search. This cleanly updates the search results.

Fixed: 1189064
Change-Id: I18f4d81af604279e8cf3c2ec0b6b245988abb2c6