[RPP] Skip interactions track screenshot tests to prevent back merge conflicts
In trying to back merge https://crrev.com/c/4506958, which fixes a
color contrast problem in the interactions track, we discovered it would
create merge conflicts because the screenshots generated in the CL would
be invalid in the relase branch (because the latter doesn't include
other changes in main that also affect the screenshots).

To avoid these conflicts, this CL temporarily disables the conflicting
tests. This CL should be cherry-picked and merged into the release
branch before the bug fix.

Once this CL lands and is cherry picked, we can enable the test again in
main.

Bug: 1442577
Change-Id: I0d9d562f322915f87714d156068506d8b39293ea