[Recorder] Move RecorderHelpers to a sub-folder
Somehow if a Node type is used within a TypeScript project, all
other Node types start to apply to the globals in unrelated files.
See the linked bug for the symptoms of this issue. A workaround
is to move modules using NodeJS types to a separate target so that
they don't pollute globals for other files. In this case, some
Puppeteer classes reference the Node types.

Fixed: 1442123
Change-Id: I735df771acdf7d4023997753f07612f79ede3784