[uma] Add instrumentation for "SourcesPanelFileDebugged" histogram
This CL records the mediaType of a UISourceCode when it is debugged.
The goal is to determine the ratio of users that either debug
sourcemapped TS/JS or debug minified JavaScript directly.

Similar to "SourcesPanelFileOpened" we record an event only once per
editor instance.

Currently we record the metric when the user:
  * Sets a breakpoint in the file
  * Pauses in the file

Since we instrument all pauses we also record stepping.

R=jarin@chromium.org

Bug: None
Change-Id: I5f4146e5caa3519b628d7d21ed0e4d3ca1d05b96