Fix DOM attribute representation in Elements panel
If an attribute value contains double-quote (") we convert
that to a single quote.
The reason is, we represent the attributes in double-quotes in Elements panel (ex: style="color: red;") which
becomes invalid syntax if the attribute value
contains a double-quote
(invalid ex: style="background: url("https://web.dev")").

This also fixes editing such attribute values in Elements panel.

Fixed: 1443236
Change-Id: I973b1a8da52963cfa4dcead79a62d3baa6eeceff