Escape double quote in attribute values while displaying them
We're marking attribute values with double quotes while displaying
them in the Elements panel.

When we're editing an attribute
and if an attribute value contains a
double quote it messes with HTML parser state
and causes the part of the attribute value to be
interpreted as an attribute name.

This CL updates those double quotes inside
attribute values to be a single quote.

Fixed: 1443236
Change-Id: I644c5f40d6e43412f8b4f4ecdbed87a28901e3cc