[RPP] Return early on tree search when selected node is a RunTask
RunTask events don't correspond to a node in the tree views and so,
if a user clicks on one of them we shouldn't look for them in the tree.

Because RunTask events span the whole time of their corresponding trees,
looking for them is expensive because all events in a tree will be
queried. Top level events that are actually included in the tree don't
have this problem because the tree is queried in preorder, so they would
be queried first.

Fixed: 1443103
Change-Id: Iaafb60c8c8177b35452e72df16563da3d533954b